requirements.txt
httpx[http2]==0.27.0
beautifulsoup4==4.12.2
lxml==4.9.3
pydantic==2.5.2
python-dateutil==2.8.2
typer==0.9.0
tqdm==4.66.1
pypdf==3.17.4
pytest==7.4.3
